feat: Refactor RegistrationRajalController to filter non-existing registration data
This commit refactors the `RegistrationRajalController` class to filter out non-existing registration data. The `index` method now retrieves data from an external API endpoint and filters out any data that does not have a corresponding entry in the `RegistrationInap` model. This change improves the accuracy of the displayed data and ensures that only valid registration data is shown to the user.